refactor(plugins/clients): decouple connectionmanager from tcp client plugin
Add separate lib for clientauthorizedprocesses and clientauthorization.
