+++ src/plugins/clients/CMakeLists.txt
@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: MIT
 
+add_subdirectory(connectionmanager)
 add_subdirectory(dummy)
 add_subdirectory(tcp)
 add_subdirectory(unix)
