+++ src/plugins/clients/connectionmanager/clientauthorizedprocesses/public/clientauthorizedprocesses/clientauthorizedprocesses.h
