+++ src/plugins/clients/connectionmanager/interface/connectionmanager/clientconnection_defines.h
