+++ src/plugins/clients/connectionmanager/interface/connectionmanager/types.h
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: MIT
+#pragma once
+
+#include "clientconnection_types.h"
+#include "connectionmanager_types.h"
