+++ src/plugins/clients/connectionmanager/private/clientblacklist.c
