+++ src/plugins/clients/connectionmanager/private/clientconnection.c
