+++ src/plugins/clients/connectionmanager/private/clientconnection_worker.c
