+++ src/plugins/clients/connectionmanager/private/connectionmanager_listen.c
@@ -6,7 +6,7 @@
 #include <sys/eventfd.h>
 #include <sys/time.h>
 
-#include "connectionmanager/clientauthorization.h"
+#include "tcp_clientauthorization/clientauthorization.h"
 #include "connectionmanager/clientconnection.h"
 #include "connectionmanager/connectionmanager.h"
 #include "connectionmanager_private.h"
@@ -16,6 +16,11 @@
 #define CONNECTION_PSELECT_TIMEOUT_SEC    0
 #define CONNECTION_PSELECT_TIMEOUT_NSEC   (100 * 1000 * 1000)
 #define TIMESPEC_1SEC_IN_NSEC             1000000000L
+
+inline bool elosUnixClientAuthorizationIsTrustedConnection(UNUSED elosClientAuthorization_t *const clientAuth,
+                                                           UNUSED struct sockaddr const *const addr) {
+    return false;
+}
 
 static inline void _calculateTvSeconds(struct timespec *ts) {
     struct timespec connSemTimeOut = {.tv_sec = CONNECTION_SEMAPHORE_TIMEOUT_SEC,
