+++ src/plugins/clients/connectionmanager/private/connectionmanager_private.h
