+++ src/plugins/clients/connectionmanager/private/connectionmanager_start.c
