+++ src/plugins/clients/connectionmanager/public/connectionmanager/clientblacklist.h
