+++ src/plugins/clients/connectionmanager/public/connectionmanager/clientconnection.h
