+++ src/plugins/clients/connectionmanager/public/connectionmanager/connectionmanager.h
