--- src/plugins/clients/tcp/interface/connectionmanager/types.h
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: MIT
-#pragma once
-
-#include "connectionmanager/clientauthorization_types.h"
-#include "connectionmanager/clientconnection_types.h"
-#include "connectionmanager/connectionmanager_types.h"
