+++ test/utest/plugins/clients/tcp/connectionmanager/clientauthorization/elosClientAuthorizationInitialize/case_err_param.c
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 
+#include <tcp_clientauthorization/clientauthorization.h>
 #include <cmocka_extensions/cmocka_extensions.h>
 #include <cmocka_mocks/mock_libmnl.h>
-#include <connectionmanager/clientauthorization.h>
 
 #include "elosClientAuthorizationInitialize_utest.h"
