+++ test/utest/plugins/clients/tcp/connectionmanager/clientauthorization/elosClientAuthorizationIsTrustedConnection/case_success.c
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 
+#include <tcp_clientauthorization/clientauthorization.h>
 #include <cmocka_mocks/mock_libmnl.h>
-#include <connectionmanager/clientauthorization.h>
 
 #include "elosClientAuthorizationIsTrustedConnection_utest.h"
