+++ test/utest/plugins/clients/tcp/connectionmanager/clientauthorizedprocesses/elosAuthorizedProcessCheck/case_success_no_match.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-#include <connectionmanager/clientauthorizedprocesses.h>
+#include <clientauthorizedprocesses/clientauthorizedprocesses.h>
 #include <elos/common/process.h>
 #include <safu/result.h>
 #include <samconf/test_utils.h>
