+++ test/utest/plugins/clients/tcp/connectionmanager/clientauthorizedprocesses/elosAuthorizedProcessDelete/case_err_authorizedprocess_param_null.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-#include <connectionmanager/clientauthorizedprocesses.h>
+#include <clientauthorizedprocesses/clientauthorizedprocesses.h>
 #include <safu/result.h>
 #include <sys/types.h>
