+++ test/utest/plugins/clients/tcp/connectionmanager/clientauthorizedprocesses/elosAuthorizedProcessDelete/case_success.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-#include <connectionmanager/clientauthorizedprocesses.h>
+#include <clientauthorizedprocesses/clientauthorizedprocesses.h>
 #include <safu/result.h>
 #include <samconf/test_utils.h>
 #include <string.h>
