+++ test/utest/plugins/clients/tcp/connectionmanager/clientauthorizedprocesses/elosAuthorizedProcessInitialize/case_success.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-#include <connectionmanager/clientauthorizedprocesses.h>
+#include <clientauthorizedprocesses/clientauthorizedprocesses.h>
 #include <samconf/test_utils.h>
 #include <string.h>
